+++ backend/run_stock.sh
@@ -1,0 +1,4 @@
+#!/bin/zsh
+
+
+/usr/bin/python3 scheduler.py