+++ changeLog.md
@@ -1,5 +1,6 @@
 ## changeLog
 
+- backend add scheduler.py,log (260617)
 - backend insert market_topic_snapshot (260615)
 - backend depoly + insert stock_main 
 - main target (260611)